Added preliminary helium letters sketch.

+++ interference_patterns/data/lux.frag
@@ -1,6 +1,6 @@
 #ifdef GL_ES
-precision mediump float;
-precision mediump int;
+precision highp float;
+precision highp int;
 #endif
 
 uniform sampler2D uTexture;
